Start supporting operations with no response
OrientDB supports three "modes" of operations:

- synchronous: the client is supposed to wait for the response
- asynchronous: the client is supposed to consume the response in a
  callback
- no-response: the client is supposed to receive no response

For the way the application is structured now, async and sync are pretty
much the same as the GenServer always performs async operations.
For now, I implemented a draft of the :no_response operation mode, which
is very simple: the GenServer replies `:ok` to the client right away,
then when it receives the response it simply parses and then ignores it.

This behaviour is a bit improvised, and I have to ask the OrientDB team
how to proceed here.