Add a GenericParser which support parsing incomplete data
This module provides a `parse/2` function that accepts a binary and a
list of functions (parsers), it returns the list of results from parsing
with each function if everything goes well, or `:incomplete` otherwise.

Each parser function must return a two element tuple with the result of
the parsing as the first element and the rest of the binary as the
second element; if parsing is unsuccessful, it must return
`:incomplete`.

This module will allow us to easily parse a response given its
specification (e.g., string + bytes + int + short + string) in a very
declarative way. If any of the response "components" is not complete,
`GenericParser` will take care of stopping the parsing and returning
`:incomplete`.

I'm not sure, but I think the need for such a complex handling of
possibly incomplete TCP messages comes from OrientDB's binary protocol
design. If, for example, each response they sent contained a session id
and then an integer with the length of the rest of the response (in
bytes), we could easily tell complete response from incomplete ones.