Add some "seeding" code to test_helper.exs
I added some code that "seeds" the OrientDB server into a known state
before running the tests (in `test_helper.exs`). This code consists of
an OrientDB SQL script that is run through the `orientdb-console`
command (that must therefore be available in `$PATH`). All these
instructions on how to run tests will be documented in the README.