Properly setup integration testing
I've been wanting to do this for a long time!

Basically, I tagged all the tests that talk to an OrientDB server with
the :integration tag. This way, `mix test` runs all the tests but `mix
test --exclude integration` runs only tests that cover stuff like
encoding and decoding of data. test_helper.exs has been changed
accordingly so that it doesn't set the database up if the tests are not
integration tests, which makes everything much much faster.

I also removed the integration testing setup I added near the beginning
of this project, since the approach introduced by this commit seems a
lot nicer to me.