Restucture handling of unknown properties
Now, records with unknown property ids are not decoded, but returned to
the caller as `MarcoPolo.UndecodedDocument` structs. If the caller finds
one of these structs in the result, it proceeds to fetch the schema
again and to decode them (on the client, not in the GenServer) with the
newly fetched schema.