Fix a bunch of failing tests (maybe due to changes in 2.2)
When a class is created with `MarcoPolo.command/2`, the response appears
not to be the cluster id for that class anymore. In fact, now classes
seem to have a bunch of cluster ids by default.

Everything works fine when using just `MarcoPolo.command/2`, but things
get tricky when using `create_record/3` passing in a cluster id
directly, as the record doesn't seem to be created.